replace flask with gunicorn in Dockerfile

+++ src/db.py
@@ -14,23 +14,19 @@ def init_app(app):
     app.cli.add_command(init_db_command)
 
 
-def init_db():
-    import models
-    db.create_all()
-
-
 @click.command("init-db")
-# @click.option("--drop", is_flag=True, help="Create after drop.")
+@click.option("--drop", is_flag=True, help="Create after drop.")
 @with_appcontext
 def init_db_command(drop):
 
     """Clear existing data and create new tables."""
-    # if drop:
-    #     click.confirm("This operation will delete the database, do you want to continue?", abort=True)
-    db.drop_all()
-    click.echo("Droped all tables.")
+    if drop:
+        click.confirm("This operation will delete the database, do you want to continue?", abort=True)
+        db.drop_all()
+        click.echo("Droped all tables.")
 
     # if define models in submodules must import them so that
     # SQLAlchemy knows about them before calling create_all
-    init_db()
+    from . import models
+    db.create_all()
     click.echo("Initialized the database.")
